Added tests for the provider and fixed bugs that came out as a result.
Of the code that was copied from riskassessment_link.py in osha.policy/patches/:
...............................................................................
Don't move the unwantedFields to the 'default' schemata, I see no
point for doing so and it complicates the test.

Also, the code that moves two fields to the bottom has been removed.

All tests now pass, although non global types are not yet created.