#999
In schema.py : Some types being extended don't have their field ordering laid out in
config.py. We however want to be certain that seoDescription will be directly
below 'description.

In config.py SPSpeaker type has extra fields specified that are not being extended.
This made the tests fail, so we remove it.